requirements.txt
# Django and web framework
Django>=5.0,<6.0
djangorestframework>=3.14.0
django-environ>=0.11.2

# Dagger SDK
dagger-io>=0.9.0

# Git operations
GitPython>=3.1.40

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
